+++ src/main/java/com/sk89q/craftbook/util/config/YAMLConfiguration.java
@@ -57,7 +57,7 @@ public class YAMLConfiguration extends LocalConfiguration {
         ICCached = config.getBoolean("circuits.ics.cache", true);
 
         config.setComment("circuits.ics.max-radius", "The max radius IC's with a radius setting can use. (WILL cause lag at higher values)");
-        ICMaxRange = config.getInt("circuits.ics.max-radius", 10);
+        ICMaxRange = config.getDouble("circuits.ics.max-radius", 10);
 
         config.setComment("circuits.ics.allow-short-hand", "Allows the usage of IC Shorthand, which is an easier way to create ICs.");
         ICShortHandEnabled = config.getBoolean("circuits.ics.allow-short-hand", true);
